added api; added auth api

+++ src/screens/Auth/SignInScreen.tsx
@@ -1,5 +1,5 @@
 import AsyncStorage from '@react-native-async-storage/async-storage';
-import { useNavigation } from '@react-navigation/core';
+import { useFocusEffect, useNavigation } from '@react-navigation/core';
 import { StackNavigationProp } from '@react-navigation/stack';
 import * as React from 'react';
 import {
@@ -8,26 +8,36 @@ import {
   Switch,
   ImageBackground,
   StatusBar,
+  ScrollView,
+  Dimensions,
 } from 'react-native';
-import { Text, Icon, Divider, Overlay } from 'react-native-elements';
+import { Text, Overlay } from 'react-native-elements';
 import { Button } from 'react-native-elements/dist/buttons/Button';
 import { KeyboardAwareScrollView } from 'react-native-keyboard-aware-scroll-view';
 import { CustomButton } from '../../components/buttons';
-import { CustomIconTextInput, CustomTextInput } from '../../components/inputs';
+import { Input } from '../../components/inputs';
 import { hooks } from '../../hooks';
 import { MainStackParamsList } from '../../navigation/types';
 import { theme } from '../../theme';
-import { wait } from '../../utils/wait';
 import * as MobX from 'mobx-react-lite';
 import { IS_ANDROID } from '../../utils/constants';
+import { FirebaseAuthErrorsCode } from '../../store/ErrorStore';
+import { useFormik } from 'formik';
+import * as Yup from 'yup';
 
 const checkIfRememberUserLogin = async () => {
   const deviceRememberMe = JSON.parse(
+    // @ts-ignore
     await AsyncStorage.getItem('remember_me'),
   );
 
   return deviceRememberMe;
 };
+
+const SignInSchema = Yup.object().shape({
+  email: Yup.string().email('Invalid email').required('Required'),
+  password: Yup.string().min(6, 'Too short').required('Required'),
+});
 
 type Navigation = StackNavigationProp<MainStackParamsList>;
 
@@ -47,11 +57,15 @@ const SignInScreen = () => {
 
   const handlePressLogin = async () => {
     try {
+      if (!(await authStore.checkIfUserExist())) {
+        formik.setErrors({
+          email: 'User is not exist',
+        });
+        return;
+      }
       setLoadinSignIn(true);
-      await authStore.fakeSignIn();
+      await authStore.signIn();
       await mainStore.getUserInstance();
-      navigation.navigate('Main', { screen: 'КАРТА' });
-
       if (rememberMe) {
         await AsyncStorage.setItem(
           'authData',
@@ -59,16 +73,51 @@ const SignInScreen = () => {
             email: authStore.email,
             password: authStore.password,
           }),
+          err => console.log('err', err),
         );
       }
+      navigation.navigate('Main', { screen: 'КАРТА' });
       !rememberMe && authStore.clear();
       setLoadinSignIn(false);
     } catch (error) {
-      console.log('eror', { error });
-
+      console.log('error', error);
+      if (error.code) {
+        const errorCode: FirebaseAuthErrorsCode = error.code;
+        switch (errorCode) {
+          case 'auth/user-not-found': {
+            formik.setErrors({ email: 'User not found' });
+            break;
+          }
+          case 'auth/wrong-password': {
+            formik.setErrors({
+              password:
+                'The password is invalid or the user does not have a password.',
+            });
+            break;
+          }
+          case 'auth/too-many-requests': {
+            formik.setErrors({
+              email:
+                'Access to this account has been temporarily disabled due to many failed login attempts. You can immediately restore it by resetting your password or you can try again later',
+            });
+            break;
+          }
+          default:
+            break;
+        }
+      }
       setLoadinSignIn(false);
     }
   };
+
+  const formik = useFormik({
+    initialValues: {
+      email: authStore.email,
+      password: authStore.password,
+    },
+    validationSchema: SignInSchema,
+    onSubmit: handlePressLogin,
+  });
 
   const handlePressForgot = () => {
     setShowForgotModal(true);
@@ -86,12 +135,17 @@ const SignInScreen = () => {
         if (result) {
           setRememberMe(result);
           const authData: { email: string; password: string } = JSON.parse(
+            //@ts-ignore
             await AsyncStorage.getItem('authData'),
           );
 
           if (authData) {
             authStore.setEmail(authData.email);
             authStore.setPassword(authData.password);
+            formik.setValues({
+              email: authData.email,
+              password: authData.password,
+            });
           }
         }
       } catch (error) {
@@ -103,72 +157,79 @@ const SignInScreen = () => {
     IS_ANDROID && StatusBar.setBackgroundColor('#000');
   }, []);
 
+  const handleChangeInput =
+    (field: 'email' | 'password') => (value: string) => {
+      formik.handleChange(field)(value);
+      switch (field) {
+        case 'email':
+          authStore.setEmail(value);
+          break;
+        case 'password':
+          authStore.setPassword(value);
+          break;
+
+        default:
+          break;
+      }
+    };
+
   return (
     <MobX.Observer
       render={() => {
         return (
           <ImageBackground
             source={require('../../../assets/BluredImage2.jpg')}
-            style={{ width: '100%', height: '100%' }}>
-            <KeyboardAwareScrollView
-              style={styles.container}
-              contentContainerStyle={{
-                width: '100%',
-                height: '100%',
-              }}>
-              <View style={styles.loginFormContainer}>
-                <View style={styles.loginForm}>
-                  <Text h1 h1Style={styles.title}>
-                    Login
-                  </Text>
-                  <CustomTextInput
-                    label={'Email'}
-                    placeholder={'Email Address'}
-                    onChangeText={authStore.setEmail}
-                    value={authStore.email}
-                  />
-                  <CustomIconTextInput
-                    label={'Password'}
-                    placeholder={'Password'}
-                    rightIcon={
-                      <Icon
-                        name="eye"
-                        type="font-awesome-5"
-                        color={'gray'}
-                        onPress={() => {}}
-                      />
-                    }
-                    secureTextEntry
-                    onChangeText={authStore.setPassword}
-                    value={authStore.password}
-                  />
-                  <View style={styles.rowContainer}>
-                    <View
-                      style={{ flexDirection: 'row', alignItems: 'center' }}>
-                      <Switch
-                        value={rememberMe}
-                        onValueChange={handleRememberMe}
-                        ios_backgroundColor={'rgb(142,142,147)'}
-                      />
-                      <Text style={styles.remember}>Remember me</Text>
-                    </View>
-                    <Button
-                      titleStyle={styles.inlineButton}
-                      title={'Forgot password?'}
-                      type={'clear'}
-                      onPress={handlePressForgot}
+            style={styles.fullSize}>
+            <ScrollView>
+              <KeyboardAwareScrollView>
+                <View style={styles.loginFormContainer}>
+                  <View style={styles.loginForm}>
+                    <Text h1 h1Style={styles.title}>
+                      Login
+                    </Text>
+                    <Input
+                      label={'Email'}
+                      placeholder={'Email Address'}
+                      onChangeText={handleChangeInput('email')}
+                      value={formik.values.email}
+                      errorMessage={formik.errors?.email}
                     />
-                  </View>
+                    <Input
+                      label={'Password'}
+                      placeholder={'Password'}
+                      secureTextEntry
+                      onChangeText={handleChangeInput('password')}
+                      value={formik.values.password}
+                      errorMessage={formik.errors?.password}
+                      secure
+                    />
+                    <View style={[styles.rowContainer, { marginBottom: 40 }]}>
+                      <View style={styles.rowContainer}>
+                        <Switch
+                          value={rememberMe}
+                          onValueChange={handleRememberMe}
+                          ios_backgroundColor={'rgb(142,142,147)'}
+                          trackColor={{ false: 'rgb(142,142,147)' }}
+                        />
+                        <Text style={styles.remember}>Remember me</Text>
+                      </View>
+                      {/* <Button
+                        titleStyle={styles.inlineButton}
+                        title={'Forgot password?'}
+                        type={'clear'}
+                        onPress={handlePressForgot}
+                      /> */}
+                    </View>
 
-                  <CustomButton
-                    title={'Login'}
-                    onPress={handlePressLogin}
-                    loading={loadingSignIn}
-                    disabled={loadingSignIn}
-                    buttonStyle={{ marginBottom: 60 }}
-                  />
+                    <CustomButton
+                      title={'Login'}
+                      onPress={formik.handleSubmit}
+                      loading={loadingSignIn}
+                      disabled={loadingSignIn}
+                      buttonStyle={{ marginBottom: 40 }}
+                    />
 
-                  {/* <View style={{ marginHorizontal: 10 }}>
+                    {/* <View style={{ marginHorizontal: 10 }}>
             <View style={{ flexDirection: 'row', alignItems: 'center' }}>
               <View style={{ height: 1, backgroundColor: '#fff', flex: 1 }} />
               <View style={{ flex: 1.5 }}>
@@ -208,34 +269,36 @@ const SignInScreen = () => {
               />
             </View>
           </View> */}
+                    <View
+                      style={[
+                        styles.rowContainer,
+                        { justifyContent: 'center' },
+                      ]}>
+                      <Text style={styles.whiteText}>
+                        Don`t have an account?
+                      </Text>
+                      <Button
+                        title="Signup"
+                        type="clear"
+                        titleStyle={styles.inlineSignUp}
+                        onPress={goToSignUp}
+                      />
+                    </View>
+                  </View>
                 </View>
-                <View style={styles.bottomSignUp}>
-                  <Text style={styles.whiteText}>Don`t have an account?</Text>
-                  <Button
-                    title="Signup"
-                    type="clear"
-                    titleStyle={styles.inlineSignUp}
-                    onPress={goToSignUp}
-                  />
-                </View>
-              </View>
-              <Overlay
-                isVisible={showForgotModal}
-                onBackdropPress={() => setShowForgotModal(false)}
-                overlayStyle={{
-                  width: '100%',
-                  marginHorizontal: 60,
-                  backgroundColor: 'gray',
-                }}>
-                {/* <View style={{ justiyContent: 'center' }}> */}
-                <CustomTextInput
-                  label={'Email'}
-                  placeholder={'Enter Email address'}
-                />
-                <CustomButton title={'Submit'} />
-                {/* </View> */}
-              </Overlay>
-            </KeyboardAwareScrollView>
+                <Overlay
+                  isVisible={showForgotModal}
+                  onBackdropPress={() => setShowForgotModal(false)}
+                  overlayStyle={{
+                    width: '100%',
+                    marginHorizontal: 60,
+                    backgroundColor: 'gray',
+                  }}>
+                  <Input label={'Email'} placeholder={'Enter Email address'} />
+                  <CustomButton title={'Submit'} />
+                </Overlay>
+              </KeyboardAwareScrollView>
+            </ScrollView>
           </ImageBackground>
         );
       }}
@@ -248,23 +311,27 @@ const styles = StyleSheet.create({
   container: {
     flex: 1,
   },
+  fullSize: {
+    width: '100%',
+    height: '100%',
+  },
   logo: {
     backgroundColor: 'purple',
   },
   loginFormContainer: {
-    height: '100%',
-    width: '100%',
+    height: Dimensions.get('screen').height,
+    width: Dimensions.get('screen').width,
     paddingHorizontal: 30,
     justifyContent: 'center',
     alignSelf: 'center',
   },
   loginForm: {
     // borderWidth: 1,
-    borderColor: '#fff',
+    // borderColor: '#fff',
   },
   title: {
     fontWeight: '600',
-    marginBottom: 80,
+    marginBottom: 50,
     textAlign: 'center',
     color: '#fff',
   },
@@ -273,7 +340,6 @@ const styles = StyleSheet.create({
     flexDirection: 'row',
     alignItems: 'center',
     justifyContent: 'space-between',
-    marginBottom: 40,
   },
   remember: {
     // color: 'rgb(142,142,147)',
